Remove appointment alerts summary HTML file
The `appointment_alerts_summary.html` file has been deleted as part of the project restructuring. This document contained a report on incomplete appointment alerts, including details such as missing appointment dates and unassigned doctors. The generation of alerts will be handled by scripts without the need for this standalone HTML file. This change aims to streamline the reporting process and reduce redundancy in documentation.